Break up pages.py into account_pages.py and calendar_pages.py
