+++ src/templates/_navigation.html
@@ -1,21 +1,21 @@
 <nav>
     <ul>
-        <li><a href="{{ url_for('pages.home') }}">Home</a></li>
-        <li><a href="{{ url_for('pages.about') }}">About</a></li>
+        <li><a href="{{ url_for('account_pages.home') }}">Home</a></li>
+        <li><a href="{{ url_for('account_pages.about') }}">About</a></li>
 
         {% if not current_user.is_authenticated %}
-            <li><a href="{{ url_for('pages.login') }}">Login</a></li>
-            <li><a href="{{ url_for('pages.signup') }}">Signup</a></li>
+            <li><a href="{{ url_for('account_pages.login') }}">Login</a></li>
+            <li><a href="{{ url_for('account_pages.signup') }}">Signup</a></li>
         {% endif %}
 
         {% if current_user.is_authenticated %}
             {% if current_user.user_type == 'doctor' %}
-                <li><a href="{{ url_for('pages.add_patient') }}">Add a Patient</a></li>
-                <li><a href="{{ url_for('pages.add_reminder') }}">Add a Reminder</a></li>
+                <li><a href="{{ url_for('calendar_pages.add_patient') }}">Add a Patient</a></li>
+                <li><a href="{{ url_for('calendar_pages.add_reminder') }}">Add a Reminder</a></li>
             {% endif %}
-            <li><a href="{{ url_for('pages.calendar') }}">Calendar</a></li>
-            <li><a href="{{ url_for('pages.profile') }}">Profile</a></li>
-            <li><a href="{{ url_for('pages.logout') }}">Logout</a></li>
+            <li><a href="{{ url_for('calendar_pages.calendar') }}">Calendar</a></li>
+            <li><a href="{{ url_for('account_pages.profile') }}">Profile</a></li>
+            <li><a href="{{ url_for('account_pages.logout') }}">Logout</a></li>
         {% endif %}
     </ul>
 </nav>
