+++ src/templates/pages/calendar.html
@@ -5,12 +5,12 @@
 {% endblock %}
 
 {% block content %}
-    <a href='{{url_for("pages.calendar", week=week - 1 ) }}'>
+    <a href='{{url_for("calendar_pages.calendar", week=week - 1 ) }}'>
         <button>
             <<
         </button>
     </a>
-    <a href='{{url_for("pages.calendar", week=week + 1 ) }}'>
+    <a href='{{url_for("calendar_pages.calendar", week=week + 1 ) }}'>
         <button>
             >>
         </button>
@@ -30,7 +30,7 @@
                 {% for reminder in reminders_by_weekday[i] %}
                     <ul class='reminder'>
                         {% if not patient %}
-                            <form action="{{ url_for('pages.delete_reminder', reminder_id=reminder.id, week=week) }}" method=post>
+                            <form action="{{ url_for('calendar_pages.delete_reminder', reminder_id=reminder.id, week=week) }}" method=post>
                                 <button>delete</button>
                             </form>
                         {% endif %}
